♻️ Moved table code to helpers
