+++ libraries/core-react/stories/components/Table.stories.tsx
@@ -13,7 +13,7 @@ import {
 } from '@components'
 import { chevron_down, chevron_up, accessible } from '@equinor/eds-icons'
 import { data, columns, Column, Data, SortDirection } from './helpers/data'
-import { toCellValues } from './hooks/toCellValues'
+import { toCellValues } from './helpers/toCellValues'
 import './styles/style.css'
 
 Icon.add({ chevron_down, chevron_up })
