+++ libraries/core-react/stories/components/Tooltip.stories.tsx
@@ -2,7 +2,7 @@ import React, { useState, useRef } from 'react'
 import styled from 'styled-components'
 import { Tooltip, TooltipProps, Typography, Button, Table } from '@components'
 import { data, columns } from './helpers/data'
-import { toCellValues } from './hooks/toCellValues'
+import { toCellValues } from './helpers/toCellValues'
 import { Story, Meta } from '@storybook/react'
 
 const Body = styled.div`
